added the diet app adn the nutritionplan model in the app and the serilizer and views .
