+++ core/urls/urls.py
@@ -29,13 +29,14 @@ from rest_framework_simplejwt.views import (
 from core.urls.user_urls import urlpatterns as users_patterns
 from core.urls.urls_membership import urlpatterns as membership_patterns
 from core.urls.urls_workout import urlpatterns as workout_patterns
-
+from core.urls.urls_diet import urlpatterns as diet_partterns
 
 urlpatterns = [
     # path('admin/', admin.site.urls),
     path("", include(users_patterns)),
     path("", include(membership_patterns)),
     path("", include(workout_patterns)),
+    path("", include(diet_partterns)),
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("schema/", SpectacularAPIView.as_view(), name="schema"),
